requirements.txt
fastapi
uvicorn
pydantic
openai
openai-agents
python-dotenv
python-multipart
